refactor sidebar to its own component
